requirementsASP.txt
functions-framework==3.*
clorm 
clingo
google-cloud-storage